+++ test/lib/src/extractor_test.dart
@@ -118,4 +118,45 @@ class UntranslatedDartFile extends StatelessWidget {
       },
     );
   });
+
+  test(
+    'setNestedValue sets value as map if it is a map and previous value not map',
+    () async {
+      // Arrange
+      final testFile = File('test/helpers/untranslated2_file.dart');
+      testFiles.add(testFile);
+
+      await testFile.parent.create(recursive: true);
+      await testFile.writeAsString(r'''
+import 'package:flutter/material.dart';
+
+class UntranslatedDartFile extends StatelessWidget {
+  const UntranslatedDartFile({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return  Column(children: [
+      Text(slang_it.untranslated.header'header text ${name}'),
+      Text(slang_it.untranslated.subheading'subheadding text ${someval.val}'),
+      Text(slang_it.untranslated.settings.heading'settings heading'),
+    ],);
+  }
+}
+
+''');
+      // Act
+      final result = await extractor.extractFromFile(testFile);
+
+      final expectedResult = {
+        'untranslated': {
+          'header': r'header text ${name}',
+          'subheading': r'subheadding text ${val}',
+          'settings': {'heading': 'settings heading'},
+        },
+      };
+      // Assert
+      expect(result, isNotEmpty);
+      expect(result, equals(expectedResult));
+    },
+  );
 }